// src/types/report.d.ts
export interface NewIncidentReport {
  // Report Data
  reportTitle: string;
  coordinates: string;
  reportDescription: string;
  severity: "LOW" | "MEDIUM" | "HIGH"; // assuming possible values
  mainType: "INCIDENT";
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  followUp: boolean;
  // Incident Data
  categoryType: IncidentCategoryType;
  incidentDescription: string;
  treatmentType: treatmentType;
  treatmentDescription: string;
  injuredBodyPart: string;

  // First Aider Details (optional)
  firstAiderName?: string;
  firstAiderPhone?: string;
  firstAiderEmail?: string;
  firstAidDate?: string;

  // Injured Person Data
  injuredPersonName: string;
  injuredPhoneNumber: string;
  injuredPersonEmail: string;
  managerSignatureConfirmationDate: string | null;
  dynamicQuestion: {
    questionId: string;
    answer: string;
  }[];
  // Media (UUIDs or URLs)
  media: string[];
}
// ...
export enum treatmentType {
  FIRST_AID = "FIRST_AID",
  NO_TREATMENT_PROVIDED = "NO_TREATMENT_PROVIDED",
  MONITORED = "MONITORED",
  DOCTOR_GP = "DOCTOR_GP",
  HOSPITAL = "HOSPITAL",
  OTHER = "OTHER",
}
export enum IncidentCategoryType {
  BUMP = "BUMP",
  CUT = "CUT",
  FRACTURE = "FRACTURE",
  RESPIRATORY = "RESPIRATORY",
  SLIP = "SLIP",
  ERGONOMIC = "ERGONOMIC",
  BREAK = "BREAK",
  PSYCHOSOCIAL = "PSYCHOSOCIAL",
  BURN = "BURN",
  FALL = "FALL",
  FAINT = "FAINT",
  CONCUSSION = "CONCUSSION",
  STRAIN = "STRAIN",
  BRUISE = "BRUISE",
  TRIP = "TRIP",
  OTHER = "OTHER",
}
export enum severity {
  LOW = "LOW",
  MEDIUM = "MEDIUM",
  HIGH = "HIGH",
  EXTREME = "EXTREME",
}